made function for repetitive blocks
