+++ app/javascript/onboarding/components/ClosingSlide.jsx
@@ -1,20 +1,11 @@
 import { h, Component } from 'preact';
 import PropTypes from 'prop-types';
 
-import { getContentOfToken } from '../utilities';
+import { stringfy } from '../utilities';
 
 class ClosingSlide extends Component {
   componentDidMount() {
-    const csrfToken = getContentOfToken('csrf-token');
-    fetch('/onboarding_update', {
-      method: 'PATCH',
-      headers: {
-        'X-CSRF-Token': csrfToken,
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({ user: { last_onboarding_page: 'closing slide' } }),
-      credentials: 'same-origin',
-    });
+    stringfy('closing slide');
   }
 
   render() {
