+++ app/javascript/onboarding/components/EmailListTermsConditionsForm.jsx
@@ -4,7 +4,7 @@ import { h, Component } from 'preact';
 import PropTypes from 'prop-types';
 
 import Navigation from './Navigation';
-import { getContentOfToken } from '../utilities';
+import { getContentOfToken, stringfy } from '../utilities';
 
 /* eslint-disable camelcase */
 
@@ -27,18 +27,7 @@ class EmailTermsConditionsForm extends Component {
   }
 
   componentDidMount() {
-    const csrfToken = getContentOfToken('csrf-token');
-    fetch('/onboarding_update', {
-      method: 'PATCH',
-      headers: {
-        'X-CSRF-Token': csrfToken,
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({
-        user: { last_onboarding_page: 'emails, COC and T&C form' },
-      }),
-      credentials: 'same-origin',
-    });
+    stringfy('emails, COC and T&C form');
   }
 
   onSubmit() {
