+++ app/javascript/onboarding/components/IntroSlide.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 
 import Navigation from './Navigation';
 import SlideContent from './SlideContent';
-import { getContentOfToken } from '../utilities';
+import { stringfy } from '../utilities';
 
 class IntroSlide extends Component {
   constructor(props) {
@@ -13,16 +13,7 @@ class IntroSlide extends Component {
   }
 
   componentDidMount() {
-    const csrfToken = getContentOfToken('csrf-token');
-    fetch('/onboarding_update', {
-      method: 'PATCH',
-      headers: {
-        'X-CSRF-Token': csrfToken,
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({ user: { last_onboarding_page: 'intro slide' } }),
-      credentials: 'same-origin',
-    });
+    stringfy('intro slide');
   }
 
   onSubmit() {
