+++ app/javascript/onboarding/utilities.js
@@ -6,3 +6,16 @@ export const jsonToForm = data => {
 
 export const getContentOfToken = token =>
   document.querySelector(`meta[name='${token}']`).content;
+
+export const stringfy = text => {
+  const csrfToken = getContentOfToken('csrf-token');
+  fetch('/onboarding_update', {
+    method: 'PATCH',
+    headers: {
+      'X-CSRF-Token': csrfToken,
+      'Content-Type': 'application/json',
+    },
+    body: JSON.stringify({ user: { last_onboarding_page: text } }),
+    credentials: 'same-origin',
+  });
+};
